move loading tables into library
